// apps/notifications-service/src/enums/notification-type.enum.ts
export enum NotificationType {
  ORDER_CREATED = 'ORDER_CREATED',
  ORDER_REQUESTED = 'ORDER_REQUESTED',
  ORDER_PAID = 'ORDER_PAID',
  ORDER_ACCEPTED = 'ORDER_ACCEPTED',
  ORDER_REJECTED = 'ORDER_REJECTED',
  ORDER_CANCELLED = 'ORDER_CANCELLED',
  ORDER_DELIVERED = 'ORDER_DELIVERED',
  ORDER_ASSIGNED = 'ORDER_ASSIGNED',

  PRODUCT_LOW_STOCK = 'PRODUCT_LOW_STOCK',
  PRODUCT_OUT_OF_STOCK = 'PRODUCT_OUT_OF_STOCK',

  PAYMENT_FAILED = 'PAYMENT_FAILED',
  REFUND_ISSUED = 'REFUND_ISSUED',

  WELCOME_USER = 'WELCOME_USER',
  PROFILE_UPDATED = 'PROFILE_UPDATED',
}